add friktion example and define cli
